Support cookie-based authentication. (#1797)
This change uses the webkit built-in CookieManager to provide cookie
policy and persistence for the okhttp client. The net effect is that
cookies that are set during the onboarding process are persisted and
used for subsequent API calls.

This allows users to secure their HomeAssistant instance using
alternate authentication mechanisms that rely on cookies
(e.g. Cloudflare Access).